Cleanup: Move uiComponents/Dialogs* into common/Modals
