+++ ui/src/components/common/AsyncTaskList/TaskWithProgressBar.tsx
@@ -1,8 +1,8 @@
 import { FC, ReactNode, useId, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { twMerge } from 'tailwind-merge';
-import { ConfirmationDialog } from '../../../uiComponents';
 import { CloseIconButton } from '../Buttons';
+import { ConfirmationDialog } from '../Modals';
 import s from './TaskWithProgressBar.module.css';
 import { ConfirmProps, Progress } from './types';
