+++ ui/src/components/map/stop-areas/EditStopAreaLayer.tsx
@@ -17,7 +17,6 @@ import {
   setSelectedMapStopAreaIdAction,
 } from '../../../redux';
 import { EnrichedStopPlace } from '../../../types';
-import { ConfirmationDialog } from '../../../uiComponents';
 import {
   getGeometryPoint,
   mapPointToStopRegistryGeoJSON,
@@ -25,6 +24,7 @@ import {
   showSuccessToast,
 } from '../../../utils';
 import { useLoader } from '../../common/hooks';
+import { ConfirmationDialog } from '../../common/Modals';
 import { StopAreaFormState, useUpsertStopArea } from '../../forms/stop-area';
 import {
   DeleteStopArea,
