+++ ui/src/uiComponents/index.ts
@@ -1,3 +1,1 @@
-export * from './ConfirmationDialog';
-export * from './DialogWithButtons';
 export * from './Switch';
